+++ app/src/app/components/contents/download/DownloadButton.tsx
@@ -1,10 +1,11 @@
-import {FileMetadata} from "../../../../types";
+import {FileMetadata, RootState} from "../../../../types";
 import {useState} from "react";
 import appConfig from "../../../../config/appConfig";
 import {download} from "../../../../lib/download";
 import {Button} from "../../Base/Button";
 import {DownloadCloud} from "lucide-react";
 import {bytesToSize} from "../../../../helpers";
+import {useSelector} from "react-redux";
 
 interface DownloadButtonProps {
     file: FileMetadata
@@ -12,14 +13,13 @@ interface DownloadButtonProps {
 
 export default function DownloadButton({ file }: DownloadButtonProps) {
     const [error, setError] = useState("");
+    const { isAuthenticated } = useSelector((state: RootState) => state.login );
 
     const downloadFile = (file: FileMetadata) => {
         const url = `${appConfig.apiUrl()}/packets/file/${file.hash}?filename=${file.path}`;
-        download(url, file.path)
-            .then(() => setError("ok now"))
+        download(url, file.path, isAuthenticated)
+            .then(() => setError(""))
             .catch(e =>  {
-                console.log("ERROR")
-                console.log(JSON.stringify(e.message))
                 setError(e.message);
             });
     };
@@ -33,7 +33,7 @@ export default function DownloadButton({ file }: DownloadButtonProps) {
                       </span>
             </Button>
             <span className="small p-2 text-muted">({bytesToSize(file.size)})</span>
-            <div className="text-red-500 h-4">{error}</div>
+            <div className="text-red-500 h-6">{error}</div>
         </>
     );
 }
