+++ app/src/lib/auth/getAuthHeader.ts
@@ -1,0 +1,6 @@
+import {getBearerToken} from "./getBearerToken";
+
+export const getAuthHeader = () => {
+    const token = getBearerToken();
+    return { Authorization: `Bearer ${token}` }
+}
