+++ app/src/lib/fetch.ts
@@ -1,4 +1,5 @@
-import { getBearerToken } from "./getBearerToken";
+import { getBearerToken } from "./auth/getBearerToken";
+import {getAuthHeader} from "./auth/getAuthHeader";
 
 export interface Fetcher {
   url: string;
@@ -9,14 +10,13 @@ export interface Fetcher {
 }
 
 export const fetcher = async ({ url, method = "GET", authRequired = false, json = true, body }: Fetcher) => {
-  const token = authRequired ? getBearerToken() : null;
   const res = await fetch(url, {
     method,
     ...(body && { body: JSON.stringify(body) }),
     headers: {
       Accept: "application/json",
       "Content-Type": "application/json",
-      ...(authRequired && { Authorization: `Bearer ${token}` })
+      ...(authRequired && getAuthHeader())
     }
   });
